16.5-> Comprehension exercise 01-> Creating `rootReducer`
