+++ front-end/16.5/comprehension-exercises/redux-tests/src/redux/reducers/clickReducer.js
@@ -1,0 +1,12 @@
+const INITIAL_STATE = {
+  counter: 0,
+};
+function clickReducer(state = INITIAL_STATE, action) {
+  switch (action.type) {
+    case 'ADD_CLICK':
+      return { counter: state.counter + 1 };
+    default:
+      return state;
+  };
+}
+export default clickReducer;
