+++ front-end/16.5/comprehension-exercises/redux-tests/src/redux/reducers/index.js
@@ -1,13 +1,6 @@
-// src/reducers/index.js
-const Initial_State = {
-  counter: 0,
-};
-function clickReducer(state = Initial_State, action) {
-  switch (action.type) {
-    case 'ADD_CLICK':
-      return { counter: state.counter + 1 };
-    default:
-      return state;
-  }
-}
-export default clickReducer;
+import { combineReducers } from 'redux';
+import clickReducer from './clickReducer';
+
+const rootReducer = combineReducers({ clickReducer });
+
+export default rootReducer;
