+++ front-end/16.5/comprehension-exercises/redux-tests/src/redux/store/index.js
@@ -1,7 +1,6 @@
 // src/store/index.js
-import { createStore, combineReducers } from 'redux';
-import clickReducer from '../reducers';
-const rootReducer = combineReducers({ clickReducer });
+import { createStore } from 'redux';
+import rootReducer from '../reducers';
 
 const store = createStore(rootReducer);
